requirements.txt
PySimpleGUI
pydantic
ffmpeg-python
requests